refactoring some render functions were moved to render-utils
